# api/openapi.yaml
openapi: 3.0.3
info:
  title: HROS Admin Service - Foundation
  version: 0.1.0
  description: Base foundation API for health and connectivity.

paths:
  /health:
    get:
      summary: Health Check
      description: Returns the health status of the application and its dependencies.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'

  /v1/auth/login:
    post:
      operationId: login
      tags:
        - Authentication
      summary: Admin Login
      description: Authenticate an admin user and return tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/auth/mfa/verify:
    post:
      operationId: mfaVerify
      tags:
        - Authentication
      summary: Verify MFA
      description: Complete MFA step for Super Admin after successful password auth.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MFAVerifyRequest'
      responses:
        '200':
          description: Verification successful, returns access/refresh tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: MFA verification failed (MFA_INVALID) or session expired (MFA_TOKEN_EXPIRED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/auth/session:
    delete:
      operationId: logout
      tags:
        - Authentication
      summary: Logout
      description: Terminate the current session by revoking the refresh token.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logout successful
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/auth/refresh:
    post:
      operationId: refresh
      tags:
        - Authentication
      summary: Refresh Access Token
      description: Refresh the access token using a valid refresh token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refresh successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/auth/password-reset/request:
    post:
      operationId: passwordResetRequest
      tags:
        - Authentication
      summary: Request Password Reset
      description: Initiate self-service password reset. Always returns success to prevent email enumeration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '200':
          description: Password reset request accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "If an account exists for that email, a reset link has been sent."
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/auth/password-reset/confirm:
    post:
      operationId: passwordResetConfirm
      tags:
        - Authentication
      summary: Confirm Password Reset
      description: Reset password using token from email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password updated successfully."
        '400':
          description: Token expired (TOKEN_EXPIRED) or already used (TOKEN_USED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Password does not meet complexity requirements (PASSWORD_WEAK)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/auth/accept-invite:
    post:
      operationId: acceptInvite
      tags:
        - Authentication
      summary: Accept Invite
      description: Complete account activation using invitation token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptInviteRequest'
      responses:
        '200':
          description: Account activated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Account activated successfully."
        '400':
          description: Invite token expired (INVITE_EXPIRED) or already used (INVITE_USED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Password does not meet complexity requirements (PASSWORD_WEAK)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/auth/sso/initiate:
    get:
      operationId: ssoInitiate
      tags:
        - Authentication
      summary: Initiate SSO
      description: Redirects the client to the Identity Provider login page.
      parameters:
        - name: provider
          in: query
          required: true
          schema:
            type: string
          example: google
      responses:
        '302':
          description: Redirect to Identity Provider login page
          headers:
            Location:
              schema:
                type: string
              description: Target URL of the Identity Provider authorization endpoint
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/auth/sso/callback:
    get:
      operationId: ssoCallback
      tags:
        - Authentication
      summary: SSO Callback
      description: Handle authentication response from the Identity Provider.
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          example: "auth_code_123"
        - name: state
          in: query
          required: true
          schema:
            type: string
          example: "state_abc"
      responses:
        '200':
          description: Login successful, returns access/refresh tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid state parameter (INVALID_SSO_STATE)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No admin account linked to this identity (NO_ACCOUNT_LINKED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/auth/biometric/challenge:
    post:
      operationId: biometricChallenge
      tags:
        - Authentication
      summary: Biometric Challenge
      description: Generate a WebAuthn challenge for biometric authentication.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BiometricChallengeRequest'
      responses:
        '200':
          description: Challenge generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BiometricChallengeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/auth/biometric/verify:
    post:
      operationId: biometricVerify
      tags:
        - Authentication
      summary: Biometric Verification
      description: Verify WebAuthn cryptographic signature and authenticate the user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BiometricVerifyRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          example: "def456..."
    HealthStatus:
      type: object
      properties:
        status:
          type: string
          example: "OK"
        version:
          type: string
          example: "0.1.0"
        dependencies:
          type: object
          additionalProperties:
            type: string
          example:
            postgres: "UP"
            redis: "UP"
            kafka: "UP"

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: admin@hros.com
        password:
          type: string
          format: password
          example: "********"
        remember_me:
          type: boolean
          example: true

    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "def456..."
        mfa_required:
          type: boolean
          example: true
        mfa_token:
          type: string
          example: "mfa_sess_abc123"
        mfa_methods:
          type: array
          items:
            type: string
          example: ["totp", "webauthn"]

    MFAVerifyRequest:
      type: object
      required:
        - mfa_token
        - method
      properties:
        mfa_token:
          type: string
          example: "mfa_sess_abc123"
        method:
          type: string
          enum: [totp, webauthn]
          example: totp
        code:
          type: string
          example: "123456"

    PasswordResetRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: admin@hros.com

    PasswordResetConfirmRequest:
      type: object
      required:
        - token
        - password
        - password_confirmation
      properties:
        token:
          type: string
          example: "uuid-reset-token-here"
        password:
          type: string
          format: password
          example: "NewSecurePass123!"
        password_confirmation:
          type: string
          format: password
          example: "NewSecurePass123!"

    AcceptInviteRequest:
      type: object
      required:
        - token
        - password
        - password_confirmation
      properties:
        token:
          type: string
          example: "uuid-invite-token-here"
        password:
          type: string
          format: password
          example: "NewSecurePass123!"
        password_confirmation:
          type: string
          format: password
          example: "NewSecurePass123!"

    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: validation_error
        message:
          type: string
          example: Request validation failed
        details:
          type: object
          additionalProperties: true
        trace_id:
          type: string
          example: 01HZX3QK0YS6M2Z1WD9S7VN9VJ

    BiometricChallengeRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: admin@hros.com

    BiometricChallengeResponse:
      type: object
      required:
        - challenge
        - credential_id
      properties:
        challenge:
          type: string
          example: "dGhpcyBpcyBhIGNyeXB0b2dyYXBoaWMgY2hhbGxlbmdl"
        credential_id:
          type: string
          example: "raw_credential_id_hex_or_base64"

    BiometricVerifyRequest:
      type: object
      required:
        - email
        - credential_id
        - authenticator_data
        - client_data_json
        - signature
      properties:
        email:
          type: string
          format: email
          example: admin@hros.com
        credential_id:
          type: string
          pattern: '^[^\s]+$'
          example: "raw_credential_id_hex_or_base64"
        authenticator_data:
          type: string
          pattern: '^[^\s]+$'
          example: "raw_authenticator_data_base64"
        client_data_json:
          type: string
          pattern: '^[^\s]+$'
          example: "raw_client_data_json_base64"
        signature:
          type: string
          pattern: '^[^\s]+$'
          example: "raw_signature_base64"
        remember_me:
          type: boolean
          example: false

  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'